requirements.txt
mlx
mlx-lm
safetensors
huggingface_hub
transformers
diffusers
torch
numpy
gguf
gradio
pillow
